Edited functionality of code
In the game_start file, I changed the import to reach the correct file. In Card, the attribute had the same name as the method, which would return an error. Lastly, created methods for first card, second card, and their guess, so that the info can be passed into do_logic.